fix(parsing): full GSI parity across sections
- wearables: parse team rosters off parsed data object not raw json
- events: accept {"event":{}} delta shape via JToken ctor
- neutralitems: parse secondary neutral1 slot
- map/minimap/courier/roshan: parse yaw float into int rotation
- generic event: value3 widened to long (uint-max sentinel survives)
- league: series_type BO1, league_node_id typo fix
- player team: new Watcher team enum (minimap lanterns)
- generic event type: smoke_activated
- 7 regression tests

+++ Dota2GSI/Nodes/CouriersProvider/Courier.cs
@@ -74,7 +74,7 @@ namespace Dota2GSI.Nodes.CouriersProvider
             IsAlive = GetBool("alive");
             RemainingRespawnTime = GetInt("respawn_time_remaining");
             Location = new Vector2D(GetInt("xpos"), GetInt("ypos"));
-            Rotation = GetInt("yaw");
+            Rotation = (int)GetFloat("yaw");
             OwnerID = GetInt("owner");
             HasFlyingUpgrade = GetBool("flying_upgrade");
             IsShielded = GetBool("shield");
